feat: add permissions for Profile and Post view sets

+++ core/views.py
@@ -1,10 +1,12 @@
 from django.http import HttpRequest, HttpResponse, HttpResponseRedirect
 from rest_framework import viewsets, mixins, serializers, status
 from rest_framework.decorators import action
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.reverse import reverse
 
 from core.models import Profile, Follow, Post, Like, Commentary
+from core.permissions import IsOwnerOrReadOnly
 from core.serializers import (
     ProfileSerializer,
     ProfileRetrieveListSerializer,
@@ -24,6 +26,12 @@ class ProfileViewSet(
 ):
     queryset = Profile.objects.all()
     serializer_class = ProfileSerializer
+    permission_classes = (IsAuthenticated,)
+
+    def get_permissions(self) -> tuple:
+        if self.action in ("update", "partial_update"):
+            return IsAuthenticated, IsOwnerOrReadOnly
+        return self.permission_classes
 
     def get_serializer_class(self) -> type(serializers.ModelSerializer):
         if self.action in ("list", "retrieve", "followings", "followers"):
@@ -36,12 +44,16 @@ class ProfileViewSet(
         url_path="me"
     )
     def my_profile(self, request: HttpRequest, *args, **kwargs) -> HttpResponse:
+        self.check_permissions(request)
+
         profile = request.user.profile
         if request.method == "GET":
             return HttpResponseRedirect(
                 reverse("social_media:profile-detail", args=[profile.id])
             )
         elif request.method == "PATCH":
+            self.check_object_permissions(request, profile)
+
             serializer = ProfileSerializer(
                 profile,
                 data=request.data,
@@ -57,6 +69,8 @@ class ProfileViewSet(
         url_path="followings"
     )
     def followings(self, request: HttpRequest, *args, **kwargs) -> HttpResponse:
+        self.check_permissions(request)
+
         followings = Follow.objects.filter(follower=request.user.profile)
         self.queryset = self.queryset.filter(
             user__in=followings.values_list("following", flat=True)
@@ -69,6 +83,8 @@ class ProfileViewSet(
         url_path="followers"
     )
     def followers(self, request: HttpRequest, *args, **kwargs) -> HttpResponse:
+        self.check_permissions(request)
+
         followers = Follow.objects.filter(following=request.user.profile)
         self.queryset = self.queryset.filter(
             user__in=followers.values("follower")
@@ -81,6 +97,8 @@ class ProfileViewSet(
         url_path="follow"
     )
     def follow(self, request: HttpRequest, *args, **kwargs) -> HttpResponse:
+        self.check_permissions(request)
+
         profile = self.get_object()
         user = request.user.profile
 
@@ -92,7 +110,10 @@ class ProfileViewSet(
         serializer = FollowSerializer(data=follow_data)
         serializer.is_valid(raise_exception=True)
 
-        relation = Follow.objects.filter(follower=user.profile, following=profile)
+        relation = Follow.objects.filter(
+            follower=user.profile,
+            following=profile
+        )
         if not relation.exists():
             serializer.save()
             return Response(
@@ -107,6 +128,14 @@ class ProfileViewSet(
 class PostViewSet(viewsets.ModelViewSet):
     queryset = Post.objects.all()
     serializer_class = PostSerializer
+    permission_classes = (IsAuthenticated,)
+
+    def get_permissions(self) -> tuple:
+        if self.action in (
+            "update", "partial_update", "destroy", "delete_comment"
+        ):
+            return IsAuthenticated, IsOwnerOrReadOnly
+        return self.permission_classes
 
     def get_serializer_class(self) -> type(serializers.ModelSerializer):
         if self.action in ("list", "liked", "my_posts", "following_posts"):
@@ -121,6 +150,8 @@ class PostViewSet(viewsets.ModelViewSet):
         url_path="my-posts"
     )
     def my_posts(self, request: HttpRequest, *args, **kwargs) -> HttpResponse:
+        self.check_permissions(request)
+
         user = request.user.profile
         self.queryset = self.queryset.filter(author=user)
         return super().list(request, *args, **kwargs)
@@ -133,6 +164,8 @@ class PostViewSet(viewsets.ModelViewSet):
     def following_posts(
         self, request: HttpRequest, *args, **kwargs
     ) -> HttpResponse:
+        self.check_permissions(request)
+
         user = request.user.profile
         followings = Follow.objects.filter(follower=user)
         self.queryset = self.queryset.filter(
@@ -146,6 +179,8 @@ class PostViewSet(viewsets.ModelViewSet):
         url_path="liked"
     )
     def liked(self, request: HttpRequest, *args, **kwargs) -> HttpResponse:
+        self.check_permissions(request)
+
         user = request.user.profile
         liked = Like.objects.filter(user=user).values_list(
             "post_id",
@@ -160,6 +195,8 @@ class PostViewSet(viewsets.ModelViewSet):
         url_path="like"
     )
     def like(self, request: HttpRequest, *args, **kwargs) -> Response:
+        self.check_permissions(request)
+
         post = self.get_object()
         user = request.user.profile
 
@@ -176,6 +213,8 @@ class PostViewSet(viewsets.ModelViewSet):
         url_path="comment"
     )
     def comment(self, request: HttpRequest, *args, **kwargs) -> HttpResponse:
+        self.check_permissions(request)
+
         post = self.get_object()
         serializer = CommentarySerializer(
             data=request.data, context={"request": request, "post": post}
@@ -198,6 +237,9 @@ class PostViewSet(viewsets.ModelViewSet):
         user = request.user.profile
         try:
             commentary = post.comments.get(id=pk_comment, author=user)
+
+            self.check_object_permissions(request, commentary)
+
             commentary.delete()
             return Response(
                 {"status": "comment deleted"},
